[develop] refactoring
